Restructure as atomic design
