+++ src/components/Molecules/PokemonTypes.js
@@ -1,14 +1,11 @@
 import React from "react";
+import PokemonType from "../Atoms/PokemonType";
 
 export default function PokemonTypes(props) {
   return (
     <div>
       {props.pokemon.types.map((type) => {
-        return (
-          <button key={props.pokemon.name + type.type.name}>
-            {type.type.name}
-          </button>
-        );
+        return <PokemonType name={props.pokemon.name} type={type.type} />;
       })}
     </div>
   );
